Update MarkTwo to run in Windows and Linux. Removed Lua support and replaced with XML.

+++ src/Core/Types/String.cpp
@@ -269,7 +269,7 @@ bool String::operator>(const char* rhs) const
 
 // -------------------------------------------------------
 // -------------------------------------------------------
-char& String::operator[](size_t index)
+char& String::operator[](uint32_t index)
 {
     return m_String[index];
 }
